pyproject.toml
[project]
name = "cell_image_SegClas"
version = "0.1.0"
description = "Cell Image Segmentation and Classification (2016)"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "numpy==1.26.4",
    "cellpose",
    "monai",
    "scipy",
    "scikit-learn",
    "pillow>=12.0.0",
    "segment-anything",
    "pandas",
    "ogb==1.3.6",
    "opencv-python",
    "opencv-python-headless",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "requests",
    "statsmodels",
    "matplotlib",
    "seaborn",
    "plotly",
    "kaleido",
    "jupyter",
    "dotenv",
    "biopython>=1.86",
    "openpyxl>=3.1.5",
    "dask>=2026.1.1",
    "decorator>=5.2.1",
    "docutils>=0.22.4",
    "entrypoints>=0.4",
    "flask>=3.1.2",
    "networkx>=3.6.1",
    "graphviz>=0.21",
    "ipykernel>=7.2.0",
    "urllib3",
    "xmltodict",
]


